Verion 1.4.1 ============ FIX: The auto updater was downloading .msi. It needs to download the .exe
Version 1.4
===========
ADD: Latest PlantUml version which supports Asynchronous Message in Sequence Diagram, Class member visibility
FIX: When the browse for folder dialog box is opened, it defaults to the path on the text box
CHANGE: The Plantuml.jar is no longer embedded inside the plantuml.exe. The exe used to extract the jar and then run and thus very slow.
CHANGE: When a file is opened for the first time and first save is done, it first creates a backup. This helps recover from accidental changes.


git-svn-id: http://plantumleditor.googlecode.com/svn/trunk@18 342e5d20-2858-11df-acdc-d92bc0439b3c